Move indirect funding to protocols folder
