+++ packages/wallet/src/__stories__/a-dummy-wallet-states.ts
@@ -1,8 +1,8 @@
-import { IndirectFundingState } from 'src/redux/indirect-funding/state';
+import { IndirectFundingState } from 'src/redux/protocols/indirect-funding/state';
 import * as fundingStates from '../redux/channel-state/funding/state';
 import * as sharedStates from '../redux/channel-state/shared/state';
 import * as channelStates from '../redux/channel-state/state';
-import * as indirectFundingPlayerA from '../redux/indirect-funding/player-a/state';
+import * as indirectFundingPlayerA from '../redux/protocols/indirect-funding/player-a/state';
 import { EMPTY_OUTBOX_STATE } from '../redux/outbox/state';
 import * as walletStates from '../redux/state';
 import { asAddress, asPrivateKey } from '../redux/__tests__/test-scenarios';
