+++ packages/wallet/src/__stories__/index.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import WalletContainer from '../containers/wallet';
 import '../index.scss';
-import * as indirectFundingPlayerA from '../redux/indirect-funding/player-a/state';
-import * as indirectFundingPlayerB from '../redux/indirect-funding/player-b/state';
+import * as indirectFundingPlayerA from '../redux/protocols/indirect-funding/player-a/state';
+import * as indirectFundingPlayerB from '../redux/protocols/indirect-funding/player-b/state';
 import * as walletStates from '../redux/state';
 import { indirectFundingWalletState as indirectFundingWalletStateA } from './a-dummy-wallet-states';
 import { indirectFundingWalletState as indirectFundingWalletStateB } from './b-dummy-wallet-states';
