+++ packages/wallet/src/redux/protocols/indirect-funding/__tests__/reducer.test.ts
@@ -1,14 +1,14 @@
-import * as actions from '../../actions';
-import * as states from '../../state';
+import * as actions from '../../../actions';
+import * as states from '../../../state';
 import { indirectFundingReducer } from '../reducer';
 import * as playerA from '../player-a/reducer';
 import * as playerB from '../player-b/reducer';
 
-import * as scenarios from '../../__tests__/test-scenarios';
-import { PlayerIndex } from '../../types';
+import * as scenarios from '../../../__tests__/test-scenarios';
+import { PlayerIndex } from '../../../types';
 import * as indirectFundingStates from '../../indirect-funding/state';
-import { EMPTY_OUTBOX_STATE } from '../../outbox/state';
-import { EMPTY_CHANNEL_STATE } from '../../channel-state/state';
+import { EMPTY_OUTBOX_STATE } from '../../../outbox/state';
+import { EMPTY_CHANNEL_STATE } from '../../../channel-state/state';
 const { channelId, ledgerChannel } = scenarios;
 const emptySharedData = {
   outboxState: EMPTY_OUTBOX_STATE,
