+++ packages/wallet/src/redux/protocols/indirect-funding/state.ts
